\name{chooseBeta}
\alias{chooseBeta}
\title{
The restart probability of β in the interaction network
}
\description{
We choose β to balance the diffusion and the retainment
}
\usage{
chooseBeta()
}
\details{
β determines the degree of diffusion, namely how far the random wanker can move in the network. 
}
\value{
%%  ~Describe the value returned
0≤β≤1
}
\author{
Ting Wei <weitinging@sjtu.edu.cn>; Zhangsheng Yu
Maintainer: Ting Wei <weitinging@sjtu.edu.cn>
}



