\name{normalizedMatrix}
\alias{normalizedMatrix}
\title{
normalizing the adjacency matrix
}
\description{
normalizing the adjacency matrix by column
}
\usage{
normalizedMatrix(A_Matrix)
}
\arguments{
  \item{A_Matrix}{the adjacency matrix of interaction network}
}
\value{
%%  ~Describe the value returned
matrix
}
\author{
Ting Wei <weitinging@sjtu.edu.cn>; Zhangsheng Yu
Maintainer: Ting Wei <weitinging@sjtu.edu.cn>
}
