\name{RankSum}
\alias{RankSum}
\title{
transforming the relevance score into ranks for each sample
}
\description{
transforming the relevance score into ranks for each sample and calculating the sum of per-sample ranks as population-level gene ranking
}
\usage{
RankSum(RelevanceScores)
}
\arguments{
  \item{RelevanceScores}{the relevance score of gene for each individual sample}
}
\value{
%%  ~Describe the value returned
matrix
}
\author{
Ting Wei <weitinging@sjtu.edu.cn>; Zhangsheng Yu
Maintainer: Ting Wei <weitinging@sjtu.edu.cn>
}
