\name{TumorNormalExp}
\alias{TumorNormalExp}
\title{
differential expression with the paired tumor and normal expression
}
\description{
we canculate the Absolute of Log2 Fold-Change (ALFC) of gene expression between the paired tumor and normal samples as differential expression
}
\usage{
TumorNormalExp(TumorExpression, NormalExpression)
}
\arguments{
  \item{TumorExpression}{the tumor sample expresion}
  \item{NormalExpression}{the normal sample expression}
}
\value{
%%  ~Describe the value returned
matrix
}
\author{
Ting Wei <weitinging@sjtu.edu.cn>; Zhangsheng Yu
Maintainer: Ting Wei <weitinging@sjtu.edu.cn>
}

